Extract post_edit_excerpt.rs from turn.rs (parent #680)
Move the Issue #636 workspace-confined, cap-bounded post-Edit/Write
excerpt reader out of `turn.rs` into a focused sibling module so
`turn.rs` keeps shrinking toward responsibility-focused units (parent
meta-issue #680).

One production helper extracted as a free function taking `&Agent`
(matching `actor_loop_flow` / `reply_retry` / earlier vertical-slice
precedent):

- `bounded_post_edit_excerpt` (path confinement + O_NOFOLLOW +
  `MAX_ARTIFACT_EXCERPT_BYTES` cap-before-read + NUL/UTF-8 guards +
  mask_secrets + mask_header_family + post-mask char-boundary
  re-truncation)

Behavior unchanged: same `resolve_user_path` + canonical-root +
`strip_prefix` path-escape rejection, same
`MAX_ARTIFACT_EXCERPT_BYTES` cap, same UTF-8/NUL guards, same
SSOT mask stacking (mask_secrets → mask_header_family →
`truncate_on_char_boundary`).

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `repo_edit_observation.rs` (1 site) and `progress_tests.rs` (15
test sites) to the free-fn form.

LOC delta:
- turn.rs: 464 → 409 (-55)
- new module: +69

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
